+++ Assignment 1/eventQueue.cpp
@@ -27,7 +27,7 @@ int main()
         case 1:
         {
             cout << "Enter coordinates of point to be inserted in the format: x y\n";
-            float x, y;
+            double x, y;
             cin >> x >> y;
             Point p = bst.createPoint(x, y);
 
@@ -39,7 +39,7 @@ int main()
         case 2:
         {
             cout << "Enter coordinates of point to be searched in the format: x y\n";
-            float x, y;
+            double x, y;
             cin >> x >> y;
             Point p = bst.createPoint(x, y);
 
@@ -58,7 +58,7 @@ int main()
         case 3:
         {
             cout << "Enter coordinates of point to be deleted in the format: x y\n";
-            float x, y;
+            double x, y;
             cin >> x >> y;
             Point p = bst.createPoint(x, y);
